\name{bpExperimentNames}
\alias{bpExperimentNames}
\docType{data}
\title{A Character Vector Containing the Names of the
	12 Graph Objects Stored in this Data Package}

\usage{data(bpGraphNames)}
\format{
  The format is:
 chr "bpGraphNames"
}
\description{
	This is a character vector of the names of the graph objects stored 
	in this data directory.
	
}

\examples{
data(bpGraphNames)
bpGraphNames
}

\keyword{datasets}