\name{bpExperimentNames}
\alias{bpExperimentNames}
\docType{data}
\title{A Character Vector Containing the Names of the
	13 Graph Objects Stored in this Data Package}

\usage{data(bpExperimentNames)}
\format{
  The format is:
 chr "bpExperimentNames"
}
\description{
	This is a character vector of the names of the graph objects stored 
	in this data directory.
	
}

\examples{
data(bpExperimentNames)
bpExperimentNames
}

\keyword{datasets}
